+++ test/externalTests/zeppelin.sh
@@ -31,7 +31,7 @@ function test_fn { npm run test; }
 function zeppelin_test
 {
     OPTIMIZER_LEVEL=1
-    setup https://github.com/OpenZeppelin/openzeppelin-solidity.git master
+    setup https://github.com/OpenZeppelin/openzeppelin-solidity.git v2.5.0
     run_install install_fn
 
     CONFIG="truffle-config.js"